feat(Form): Dynamically serialize string & numbers
Currently, `Number` schemas are serialized as `String` due to both
schemas using `StringField`.

This commit dynamically choose how to serialize the value, depending on
the schema type and the actual value.

As an example, for a `Number` schema, numbers will be serialized as
numbers, while placedholders like `{{ property }}` will be serialized as
text.

fix: https://github.com/KaotoIO/kaoto/issues/1653